Rename loan list related components
LoanList -> UserLoans
LoansDetails -> QuotasSection
LoanListSlider -> LoanSlider
